bug fix in generate participants and print status for floating point numbers

+++ include/initializer.hpp
@@ -20,7 +20,7 @@ public:
 
 private:
     Participant get_random_participant(std::string name);
-    Participant get_participant( std::string name, int coin_value);
+    Participant get_participant( std::string name, double coin_value);
     Coin get_random_coin(int interval);
 };
